Add files via upload
The game's code. Simple button game where the buttons act as switches with LED's. All of the switches are broken when not pressed, and when a full circuit is relayed back to the Arduino, it picks it up and translates it to a button being pressed, and if the game is active during that press, it gets logged as a point, the button pressed gets lit down and 'timed out' in order to avoid multiple points being given to a single press, and then one of the other 8 buttons gets lit up as the target button by random choice, and the loop repeats for 30 seconds, with the goal of getting the highest score being the point of this game. The button, thankfully, already had a resistor built in, so I was able to avoid having to put unneeded electronic components into the breadboards.